% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmean.R
\name{ccmean}
\alias{ccmean}
\title{Calculates estimates of mean cost with censored data}
\usage{
ccmean(x, id = "id", cost = "cost", start = "start", stop = "stop",
  delta = "delta", surv = "surv", L = NA, addInterPol = 0)
}
\arguments{
\item{x}{A dataframe with columns id, cost, start, stop, delta, surv. If columns are named differently use following parameters to specify them.}

\item{id}{The id separating each individual}

\item{cost}{The total cost, or if start and stop provided the specific cost}

\item{start}{Start of cost}

\item{stop}{End of cost, if one time cost then start = stop}

\item{delta}{Event variable, 1 = event, 0 = no event}

\item{surv}{Survival}

\item{L}{Limit. Mean cost is calculated up until L.}

\item{addInterPol}{This parameter affects the interpolation of cost between two observed times. Defaults to zero.}
}
\value{
An object of class "ccobject".
}
\description{
This function calcutes the mean cost for right-censored cost 
data over a period of L time units (days, months, years,...)
}
\details{
The function returns 4 estimates. The first two are simple and biased
downwards, and included for comparisson. The estimates are:

- AS: "Available Sample" - The simple sample mean

- CC: "Complete Case" - The mean of fully observed cases

- BT: "Weighted Complete Case" - Bang and Tsiatis's method 

- ZT: "Weighted Available Sample" - Zhao and Tian's method
}
\examples{
hcost
ccmean(hcost, L = 1461, addInterPol = 1)

}
\references{
\insertRef{Bang2000}{ccostr}

\insertRef{Zhao2001}{ccostr}
}
