% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcost.R
\docType{data}
\name{hcost}
\alias{hcost}
\title{Simulated data from the stata hcost package}
\format{An object of Rdata}
\source{
\href{http://shuaichen.weebly.com/developed-package.html}{Blog}
}
\usage{
data(hcost)
}
\description{
Simulated data from the stata hcost package
}
\examples{
data(hcost)


}
\references{
\insertRef{Chen2015}{ccostr}
}
\keyword{datasets}
