% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot.ccobject}
\alias{plot.ccobject}
\title{Adding to the generic plot function}
\usage{
\method{plot}{ccobject}(obj)
}
\arguments{
\item{plot.ccobject}{Plots the results}
}
\description{
Adding to the generic plot function
}
\examples{

}
