% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{print.ccobject}
\alias{print.ccobject}
\title{Adding to the generic print function}
\usage{
\method{print}{ccobject}(obj)
}
\arguments{
\item{print.ccobject}{Returns a nicely formatted output}
}
\description{
Adding to the generic print function
}
\examples{

}
