% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim.patients}
\alias{sim.patients}
\title{Simulates data for control of estimates of censored costs - MARKOW}
\usage{
sim.patients(n, censor = 0.05, cost)
}
\arguments{
\item{sim.patients}{Simulates censored data}
}
\value{
Different simulations
}
\description{
Transition rate matrix for the illness-death model with recovery
State 1: No treatment; State 2: Treatment; State 3: Death; State 4: Censoring
}
\examples{
sim.patients(n = 100, censor = 0.05, cost = 100)
}
