% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{simCostData}
\alias{simCostData}
\title{Simulates data for control of estimates of censored costs - LIN}
\usage{
simCostData(n = 100, dist = "unif", censor = "light", L = 10)
}
\arguments{
\item{simCostData}{Simulates censored data}
}
\value{
Different simulations
}
\description{
Simulates data for control of estimates of censored costs - LIN
}
\examples{
simCostData(n = 100, dist = "unif", censor = "light", L = 10)
}
