% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmean.R
\name{ccmean}
\alias{ccmean}
\title{Calculates estimates of mean valus given censored cost data}
\usage{
ccmean(x, id = "id", tcost = "tcost", delta = "delta",
  surv = "surv")
}
\arguments{
\item{ccmean}{Converts the questionare scores}
}
\value{
Different estimates
}
\description{
This function converts scores from the EQ-5D-5L questionaire to Quality Adjusted Life Years.
}
\details{
- Naive "full sample"
- Naive "complete case"
- Lin's method
- Bang and Tsiatis's method
}
\examples{
ccmean(df, id="id", tcost="tcost", delta="delta", surv="surv")
}
