% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmean.R
\name{ccmean}
\alias{ccmean}
\title{Calculates estimates of mean valus given censored cost data}
\usage{
ccmean(x, id = "id", cost = "cost", start = "start", stop = "stop",
  delta = "delta", surv = "surv")
}
\arguments{
\item{ccmean}{Caldulates the estimates of mean costs}
}
\value{
Different estimates
}
\description{
This function converts scores from the EQ-5D-5L questionaire to Quality Adjusted Life Years.
}
\details{
- Naive "Available Sample"
- Naive "Complete Case"

- LinT - Lin's method
- BT - Bang and Tsiatis's method 
- ZT - Zhao and Tian's method
}
\examples{
ccmean(df, id="id", tcost="tcost", delta="delta", surv="surv")
}
