% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccmean.R
\name{ccmean}
\alias{ccmean}
\title{Calculates estimates of mean valus given censored cost data}
\usage{
ccmean(x, id = "id", cost = "cost", start = "start", stop = "stop",
  delta = "delta", surv = "surv", L = NA, addInterPol = 0)
}
\arguments{
\item{ccmean}{Calculates the estimates of mean costs}
}
\value{
Mean, Variance, SD, and 95% CI for the different estimates
}
\description{
This function calcutes the mean cost for right-censored cost data over a period of L time units (days, months, years,...)
}
\details{
- Naive "Available Sample"
- Naive "Complete Case"

- LinT - Lin's method
- BT - Bang and Tsiatis's method 
- ZT - Zhao and Tian's method
}
\examples{
ccmean(df, id="id", tcost="tcost", delta="delta", surv="surv", L = NA)
}
