% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hcost.R
\docType{data}
\name{hcost}
\alias{hcost}
\title{Simulated data from the stata hcost package}
\format{
A data frame with 9882 rows and 7 variables:
\describe{
  \item{id}{id seperating individuals}
  \item{start}{start of specified cost}
  \item{stop}{end of specified cost}
  \item{cost}{cost in given period}
  \item{trt}{treatment variable}
  \item{delta}{event variable, 0 = censored}
  \item{surv}{survival period}
}
}
\source{
\href{https://shuaichen.weebly.com/developed-software.html}{Blog}
}
\usage{
data(hcost)
}
\description{
Simulated data from the stata hcost package
}
\examples{
data(hcost)


}
\references{
\insertRef{Chen2015}{ccostr}
}
\keyword{datasets}
