% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_functions.R
\name{plot.ccobject}
\alias{plot.ccobject}
\title{Adding to the generic plot function}
\usage{
\method{plot}{ccobject}(x, ...)
}
\arguments{
\item{x}{The ccobject}

\item{...}{passthrough}
}
\value{
a plot
}
\description{
Adding to the generic plot function
}
