% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{simCostData}
\alias{simCostData}
\title{Simulates censored cost data}
\usage{
simCostData(n = 100, dist = "unif", censor = "light", L = 10)
}
\arguments{
\item{n}{Number of individuals to simulate}

\item{dist}{Survival distribution either "unif" = unif(0,10) o r "exp" = exp (1/6)}

\item{censor}{Censoring "light" or "heavy" for unif(0,20) or unif(0,12.5)}

\item{L}{Number of years to summarize over}
}
\value{
Different simulations
}
\description{
Simulates censored cost data
}
\examples{
# The simulated data can be used to show how the estimators perform

simCostData(n = 100, dist = "unif", censor = "light", L = 10)
}
\references{
\insertRef{Lin1997}{ccostr}
}
